% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_facet_cleaning.R
\name{add_facets_manually}
\alias{add_facets_manually}
\title{Add Facets Interactively}
\usage{
add_facets_manually(
  clusters,
  local_heights,
  selection.indices.add,
  plot_diam_facets = 50,
  plot_diam_points = plot_diam_facets/10
)
}
\arguments{
\item{clusters}{A tibble containing facet coordinates in columns \verb{x, y, z}.}

\item{local_heights}{A tibble containing all vertex coordinates extracted
from STLs in columns \verb{x, y, z}.}

\item{selection.indices.add}{A numerical list of facet indices belonging to
to \code{local_heights} to add to \code{clusters}.
In the first run of \code{add_facets_manually()}, these should be set to
\code{NULL}.The list will then be amended with every iteration of
\code{add_facets_manually()}. Make sure to not reset this list, because than
all info on selected facets to be removed will be gone.}

\item{plot_diam_facets}{A numeric value for plotting facets. This should be
roughly the size of one facet. Default: \code{50}.}

\item{plot_diam_points}{A numeric value for plotting points. This should be
roughly 1/10th of the size of one facet. Default: \code{plot_diam_facets/10}.}
}
\value{
\code{selection.indices.add} See above.
}
\description{
Add facets that that are missing interactively.
}
\examples{
# xxx: add example

}
