% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_facet_peaks_fine}
\alias{find_facet_peaks_fine}
\title{Fine clustering towards Facet Peaks}
\usage{
find_facet_peaks_fine(df, cols_to_use)
}
\arguments{
\item{df}{A tibble containing triangle center coordinates of vertices that
lie above threshold in columns \verb{x, y, z} .
Typically, this is the resulting tibble of the \code{find_facet_peaks_rough()}
function.}

\item{cols_to_use}{A numerical value of which columns store x,y,z coordinates
of rough clustering.}
}
\value{
Tibble \code{df} with ideally one vertex per facet, located on the facet
peak.
}
\description{
Find facet peaks.
}
\examples{
# xxx: add example

}
