% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_checkpoint.R
\name{read_checkpoint}
\alias{read_checkpoint}
\title{Import Checkpoint Landmarks}
\usage{
read_checkpoint(
  folder,
  pattern = "ckpt",
  recursive = TRUE,
  keep_missing = TRUE
)
}
\arguments{
\item{folder}{A character string with the full path a folder with
Stratovan Checkpoint files.}

\item{pattern}{A character string defining what Checkpoint files to look for.}

\item{recursive}{A logical value to define if subfolders should be searched
as well. Defailt: \code{TRUE}.}

\item{keep_missing}{A logical value defining if landmarks declared as missing
should be kept. Defailt: \code{TRUE}.}
}
\value{
A dataframe with defined the columns \verb{X, Y, Z,} = landmark
coordinates, \code{LM} = landmark names, and \code{file_name} = name of file where
landmark was extracted from.
}
\description{
Imports landmarks from Stratovan Checkpoint files (https://www.stratovan.com/).
}
\examples{
# xxx: add example

}
