% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_facet_cleaning.R
\name{remove_facets_manually}
\alias{remove_facets_manually}
\title{Remove Facets Interactively}
\usage{
remove_facets_manually(
  clusters,
  local_heights,
  selection.indices.rem,
  plot_diam_facets = 50,
  plot_diam_points = plot_diam_facets/10
)
}
\arguments{
\item{clusters}{A tibble containing facet coordinates in columns \verb{x, y, z}.}

\item{local_heights}{A tibble containing all vertex coordinates extracted
from STLs in columns \verb{x, y, z}.}

\item{selection.indices.rem}{A numerical list of facet indices to remove.
In the first run of \code{remove_facets_manually()}, these should be set to
\code{NULL}.The list will then be amended with every iteration of
\code{remove_facets_manually()}. Make sure to not reset this list, because than
all info on selected facets to be removed will be gone.}

\item{plot_diam_facets}{A numeric value for plotting facets. This should be
roughly the size of one facet. Default: \code{50}.}

\item{plot_diam_points}{A numeric value for plotting points. This should be
roughly 1/10th of the size of one facet. Default: \code{plot_diam_facets/10}.}
}
\value{
\code{selection.indices.rem} See above.
}
\description{
Remove facets that have been identified by the \code{find_facet_peaks_fine()}
function interactively.
}
\examples{
# xxx: add example

}
