% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_to_R.R
\name{STL_2_tibble}
\alias{STL_2_tibble}
\title{Import STL as Tibble}
\usage{
STL_2_tibble(file_name)
}
\arguments{
\item{file_name}{File name of STL to import.}
}
\value{
A tibble containing triangle centers and triangle normals of STL file.
}
\description{
Imports triangle centers and triangle normals of STL file as tibble.
}
\examples{
# xxx: add example

}
